organized a few files and tried to get remote selection working.
