+++ 3380 Project/src/components/Category Creation/InProgressColumn.tsx
@@ -9,7 +9,7 @@ import {
   deleteBackendTask,
   fetchTasks,
   updateBackendTask,
-} from "../../firebase/CategoryCreationCrudFunctions";
+} from "../../firebase/CRUD/CategoryCreationCrudFunctions";
 
 const InProgressColumn: React.FC = () => {
   const [tasks, setTasks] = useState<Task[]>([]);
