+++ 3380 Project/src/firebase/CRUD/RemoteSelectionCRUD.ts
@@ -1,4 +1,4 @@
-import { database } from "./firebase";
+import { database } from "../firebase";
 import { doc, deleteDoc, collection, addDoc, getDocs } from "firebase/firestore";
