+++ 3380 Project/src/firebase/CRUD/SubjectTabsCrudFunctions.tsx
@@ -1,4 +1,4 @@
-import { database } from "./firebase";
+import { database } from "../firebase";
 import {
   collection,
   addDoc,
